Specify version (3.8.1) of maven-compiler-plugin
It's never a good idea not to specify the version, because then it
refers to the latest version that could break the build process when a
new version is released.

[WARNING] Some problems were encountered while building the effective
model for net.badlion:blcmodapivelocity:jar:1.2.0
[WARNING] 'build.plugins.plugin.version' for
org.apache.maven.plugins:maven-compiler-plugin is missing. @ line 24,
column 21
[WARNING]
[WARNING] It is highly recommended to fix these problems because they
threaten the stability of your build.
[WARNING]
[WARNING] For this reason, future Maven versions might no longer support
building such malformed projects.